Add geometry PGM files
Added the files that contain the inner cell geometries which should be
simulated (the magic number was changed to P3 from the original P2
because apparently the function requires a 3, even if that was not the
original format. In this case a value of 255 means fluid cell and a
value of 1 means obstacle cell. This should be defined in a header file.