boundary_values updated and fix in main in vtk_output (name)
Extend the function boundary values in such a way that, for each initial
distribution of boundary value parameters wl, wr, wt and wb, with the
help of if or switch command, the correct boundary values for u and v
are set. Also changed boundary_val.h to reflect the change in input and
main.c to change arguments.