Implementation of init_flag
+Changed helper.c read_pgm file to assign 1 to borders (currently not
used but just in case)
+Added macros in helper.h for obstacle cells and boundary cell
classification used in the flag field.
+Created the function that initializes the whole flag field reading the
pgm file and assigning values according to the neighbouring fluid cells.
I also added
a warning message in case there are more than 2 neighbours, but we could
put a full handler instead of just the message
+Added the function definition to init.h
+ In main.c, added the flag matrix and initialized it as imatrix, also
freed up memory in the end and call the function to init the flag field.