Updated SOR.C function
***CHECK IF THE BOUNDARY PRESSURE VALUES SHOULD BE SET THERE OR BEFORE
OR AFTER********************
In SOR, apart from the boundary values for the pressure at the boundary
stripe, also the boundary values in the interior boundary cells should
be set before each iteration step.
In SOR, the iteration and the residual computation needs to be limited
to the fluid cells. Besides, the normalization of the residual in SOR
should be produced by dividing by the current number of fluid cells, and
not by imax  jmax.